Refakturert opprettelse og bruk av BronnoysundClient
