+++ src/main/kotlin/no/nav/amt_enhetsregister/client/Moderenhet.kt
@@ -1,0 +1,7 @@
+package no.nav.amt_enhetsregister.client
+
+data class Moderenhet(
+	val organisasjonsnummer: String,
+	val navn: String,
+	val slettedato: String?,
+)
