+++ src/main/kotlin/no/nav/amt_enhetsregister/client/Underenhet.kt
@@ -1,0 +1,8 @@
+package no.nav.amt_enhetsregister.client
+
+data class Underenhet(
+	val organisasjonsnummer: String,
+	val navn: String,
+	val slettedato: String?,
+	val overordnetEnhet: String?,
+)
