+++ src/main/resources/application.properties
@@ -17,3 +17,4 @@ spring.datasource.password=${DB_PASSWORD}
 spring.datasource.hikari.maximum-pool-size=4
 
 app.env.virksomheterTopic=${VIRKSOMHETER_TOPIC:#{null}}
+app.env.brreg-url=https://data.brreg.no
