+++ src/test/kotlin/no/nav/amt_enhetsregister/client/BronnoysundClientTest.kt
@@ -1,25 +1,43 @@
 package no.nav.amt_enhetsregister.client
 
+import com.fasterxml.jackson.databind.ObjectMapper
 import no.nav.amt_enhetsregister.test_utils.ResourceUtils
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
-import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions.assertNotNull
+import org.junit.jupiter.api.Assertions.assertNull
+import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.autoconfigure.json.JsonTest
 import java.time.LocalDateTime
 import java.time.ZoneOffset
 import java.time.ZonedDateTime
 
-class BronnoysundClientImplTest {
+@JsonTest
+class BronnoysundClientTest {
+
+	@Autowired
+	private lateinit var objectMapper: ObjectMapper
+
+	private lateinit var server: MockWebServer
+	private lateinit var serverUrl: String
+	private lateinit var client: BronnoysundClient
+
+	@BeforeEach
+	fun setup() {
+		server = MockWebServer()
+		serverUrl = server.url("/").toString().removeSuffix("/")
+
+		client = BronnoysundClient(
+			bronnoysundUrl = serverUrl,
+			objectMapper = objectMapper,
+		)
+	}
 
 	@Test
 	fun `hentModerenhet skal lage riktig request og parse response`() {
-		val server = MockWebServer()
-		val serverUrl = server.url("").toString().dropLast(1) // Removes trailing "/"
-
-		val client = BronnoysundClientImpl(
-			bronnoysundUrl = serverUrl
-		)
-
 		val responseBody = ResourceUtils.getResourceAsText("/client/hent-moderenhet-response.json")
 
 		server.enqueue(MockResponse().setBody(responseBody))
@@ -37,13 +55,6 @@ class BronnoysundClientImplTest {
 
 	@Test
 	fun `hentModerenhet skal returnere null for status 410`() {
-		val server = MockWebServer()
-		val serverUrl = server.url("").toString().dropLast(1) // Removes trailing "/"
-
-		val client = BronnoysundClientImpl(
-			bronnoysundUrl = serverUrl
-		)
-
 		server.enqueue(MockResponse().setResponseCode(410))
 
 		assertNull(client.hentModerenhet("12345678"))
@@ -51,13 +62,6 @@ class BronnoysundClientImplTest {
 
 	@Test
 	fun `hentUnderenhet skal lage riktig request og parse response`() {
-		val server = MockWebServer()
-		val serverUrl = server.url("").toString().dropLast(1) // Removes trailing "/"
-
-		val client = BronnoysundClientImpl(
-			bronnoysundUrl = serverUrl
-		)
-
 		val responseBody = ResourceUtils.getResourceAsText("/client/hent-underenhet-response.json")
 
 		server.enqueue(MockResponse().setBody(responseBody))
@@ -76,13 +80,6 @@ class BronnoysundClientImplTest {
 
 	@Test
 	fun `hentUnderenhet skal returnere null for status 410`() {
-		val server = MockWebServer()
-		val serverUrl = server.url("").toString().dropLast(1) // Removes trailing "/"
-
-		val client = BronnoysundClientImpl(
-			bronnoysundUrl = serverUrl
-		)
-
 		server.enqueue(MockResponse().setResponseCode(410))
 
 		assertNull(client.hentUnderenhet("12345678"))
@@ -90,13 +87,6 @@ class BronnoysundClientImplTest {
 
 	@Test
 	fun `hentModerenhetOppdateringer skal lage riktig request og parse response`() {
-		val server = MockWebServer()
-		val serverUrl = server.url("").toString().dropLast(1) // Removes trailing "/"
-
-		val client = BronnoysundClientImpl(
-			bronnoysundUrl = serverUrl
-		)
-
 		val responseBody = ResourceUtils.getResourceAsText("/client/hent-moderenhet-oppdateringer-response.json")
 
 		server.enqueue(MockResponse().setBody(responseBody))
@@ -105,7 +95,10 @@ class BronnoysundClientImplTest {
 
 		val request = server.takeRequest()
 
-		assertEquals("$serverUrl/enhetsregisteret/api/oppdateringer/enheter?oppdateringsid=10&size=100", request.requestUrl.toString())
+		assertEquals(
+			"$serverUrl/enhetsregisteret/api/oppdateringer/enheter?oppdateringsid=10&size=100",
+			request.requestUrl.toString()
+		)
 
 		assertEquals(2, oppdateringer.size)
 
@@ -127,13 +120,6 @@ class BronnoysundClientImplTest {
 
 	@Test
 	fun `hentUnderenhetOppdateringer skal lage riktig request og parse response`() {
-		val server = MockWebServer()
-		val serverUrl = server.url("").toString().dropLast(1) // Removes trailing "/"
-
-		val client = BronnoysundClientImpl(
-			bronnoysundUrl = serverUrl
-		)
-
 		val responseBody = ResourceUtils.getResourceAsText("/client/hent-underenhet-oppdateringer-response.json")
 
 		server.enqueue(MockResponse().setBody(responseBody))
@@ -142,7 +128,10 @@ class BronnoysundClientImplTest {
 
 		val request = server.takeRequest()
 
-		assertEquals("$serverUrl/enhetsregisteret/api/oppdateringer/underenheter?oppdateringsid=10&size=100", request.requestUrl.toString())
+		assertEquals(
+			"$serverUrl/enhetsregisteret/api/oppdateringer/underenheter?oppdateringsid=10&size=100",
+			request.requestUrl.toString()
+		)
 
 		assertEquals(2, oppdateringer.size)
 
@@ -161,5 +150,4 @@ class BronnoysundClientImplTest {
 		assertEquals("88445533", oppdateringer[1].organisasjonsnummer)
 		assertEquals(EnhetOppdateringType.NY, oppdateringer[1].endringstype)
 	}
-
 }
